#147 - move shared components from /components/shared to /components
