+++ client/src/components/Characters/components/CharacterLabelForms/2_CharacterDetailsPathFinder_1_0_0.js
@@ -6,7 +6,7 @@ import { Form, Grid, Segment } from 'semantic-ui-react';
 import Yup from 'yup';
 
 import { propsChanged, propsBase } from 'services/props';
-import FormFieldErrorMessage from 'components/shared/components/FormFieldErrorMessage';
+import FormFieldErrorMessage from 'components/shared/FormFieldErrorMessage';
 
 
 export default class CharacterDetails5e_1_0_0 extends Component {
