+++ client/src/components/shared/CharactersSelect/index.js
@@ -7,8 +7,8 @@ import Select from 'react-select';
 import { compose } from 'recompose';
 import { Form, Image, Item, Label, Message } from 'semantic-ui-react';
 
-import { availableCharactersQuery } from '../../../Characters/queries';
-import { getFullImageUrl } from '../../../../services/image';
+import { availableCharactersQuery } from 'components/Characters/queries';
+import { getFullImageUrl } from 'services/image';
 
 class CharactersSelect extends Component {
